Add NIfTI segmentation overlay support to DICOMweb Gateway
- Introduced optional NIfTI segmentation table configuration in `app.yml` and `app-config.yml`.
- Implemented new API endpoints for fetching NIfTI overlay metadata and files.
- Updated handlers and query builders to support NIfTI overlay functionality.
- Enhanced documentation in the deployment notebook regarding NIfTI integration.
