+++ apps/dicom-web-gateway/app-config.yml
@@ -4,6 +4,8 @@ env:
     valueFrom: "sql_warehouse"
   - name: "DATABRICKS_PIXELS_TABLE"
     value: "{PIXELS_TABLE}"
+  - name: "NIFTI_SEGMENTATION_TABLE"
+    value: "{NIFTI_SEGMENTATION_TABLE}"
 
   # ── Authentication ──────────────────────────────────────────────────────────
   - name: "DICOMWEB_USE_USER_AUTH"
