+++ apps/dicom-web/ohif/index.html
@@ -3,4 +3,4 @@
         return import(moduleId);
       }
 
-      window.PUBLIC_URL = './';</script><script defer="defer" src="./app.bundle.13cca4f1c7f4b7808e38.js"></script><link href="./app.bundle.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="react-portal"></div><div id="root"></div></body></html>
+      window.PUBLIC_URL = './';</script><script defer="defer" src="./app.bundle.6bf41afda50f7ddf6348.js"></script><link href="./app.bundle.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="react-portal"></div><div id="root"></div></body></html>
