+++ apps/dicom-web/ohif/sw.js
@@ -50,7 +50,7 @@ self.addEventListener('message', event => {
     }
   }
 });
-workbox.precaching.precacheAndRoute([{'revision':null,'url':'./1398.bundle.667fbe38ca6d9fdfea19.js'},{'revision':null,'url':'./1403.bundle.75d8ae601fad37ac59b9.js'},{'revision':null,'url':'./1459.bundle.0df00d11af8704537c76.js'},{'revision':null,'url':'./147.bundle.037f190998936d61183b.js'},{'revision':'e0cfcadcf2c2fbc1f117196b9f62bf3e','url':'./147.css'},{'revision':null,'url':'./1554.bundle.313701fc6b7c2b153c28.js'},{'revision':null,'url':'./1608.bundle.7879b1f7ca74b87faf8c.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./1608.css'},{'revision':null,'url':'./1730.bundle.9a528d67a7ddfefd4ac8.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./1730.css'},{'revision':null,'url':'./1758.bundle.9574b448d811268c73b4.js'},{'revision':null,'url':'./17dd54813d5acc10bf8f.wasm'},{'revision':null,'url':'./1927.bundle.da05087c08bf79ab6164.js'},{'revision':null,'url':'./1933.bundle.f47a3916c63908ce4510.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./1933.css'},{'revision':null,'url':'./2018.bundle.a75601d36f8335e6ad70.js'},{'revision':null,'url':'./2068.bundle.bbd5a91f3a9f27e42002.js'},{'revision':null,'url':'./213.bundle.4f5a9f54de1d8b00dbe0.js'},{'revision':null,'url':'./2199.bundle.05379e24d0559fb1b62d.js'},{'revision':null,'url':'./2424.bundle.200b0ae6a4293279fd9b.js'},{'revision':null,'url':'./2516.bundle.299b43d001321aeb1554.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./2591.css'},{'revision':null,'url':'./2701.bundle.57fa57f3211b5af30333.js'},{'revision':null,'url':'./2764.bundle.eeb852510f3006df2e59.js'},{'revision':null,'url':'./2816.bundle.838fa4d72921c026083c.js'},{'revision':null,'url':'./2838.bundle.8d4f0cf2c02681ca6878.js'},{'revision':null,'url':'./31fb9346313fc3740d7b.woff2'},{'revision':null,'url':'./3343.bundle.a4424b8418b8c0c1a9e4.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./3343.css'},{'revision':null,'url':'./3461.bundle.58547d3ab516df762f48.js'},{'revision':null,'url':'./3562.bundle.6b7411fbb94c93a34820.js'},{'revision':null,'url':'./3910.bundle.b6180cdfa45c750c206b.js'},{'revision':'2b8874726752c2c40347df5b55a1e4f9','url':'./3910.css'},{'revision':null,'url':'./3933.bundle.aadc7dabcb74e4ca266c.js'},{'revision':'afd04227d29073ce2d3c1662140270d7','url':'./3933.css'},{'revision':null,'url':'./4202.bundle.be22ef13c3844ef0f8c4.js'},{'revision':null,'url':'./4287.bundle.87e14e648f2bce5e4176.js'},{'revision':null,'url':'./4775.bundle.d0ea4b37875cfa85e56c.js'},{'revision':null,'url':'./4819.bundle.f7757d944c785f953b6b.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./4819.css'},{'revision':null,'url':'./5004fdc02f329ce53b69.wasm'},{'revision':null,'url':'./5028.bundle.cd59aeb28444f14ad1ee.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./5028.css'},{'revision':null,'url':'./5261.bundle.2655560097e9250eac44.js'},{'revision':null,'url':'./5299.bundle.4ae3f0d5a95b5483f5e9.js'},{'revision':null,'url':'./5457.bundle.c44221459be9b9c5937c.js'},{'revision':null,'url':'./5485.bundle.1add1e84bf0652e1f978.js'},{'revision':null,'url':'./5491.bundle.4866d2ecb20dd089e071.js'},{'revision':null,'url':'./5547.bundle.10a9625eac35067f6c77.js'},{'revision':'7b35665742c835684bc86d3fa5397b30','url':'./5547.css'},{'revision':null,'url':'./5609.bundle.41430e142f0b95cb32e9.js'},{'revision':null,'url':'./5724.bundle.62ea5c07c7809ab4cd74.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./5724.css'},{'revision':null,'url':'./5802.bundle.d0b1eda248c9ce5ff880.js'},{'revision':'78c6e381862b3f33316bcffadde4ccd6','url':'./5802.css'},{'revision':null,'url':'./5830.bundle.50dcea7ba7954661d2b7.js'},{'revision':null,'url':'./5858.bundle.466e58128de344ab53f3.js'},{'revision':null,'url':'./6027.bundle.4b70caf6273a7926016d.js'},{'revision':null,'url':'./6163.bundle.09921650f5c2a777c59a.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./6163.css'},{'revision':null,'url':'./6280.bundle.f22f9955eb906e1d7da8.js'},{'revision':null,'url':'./6347.bundle.46b118d368710e08c546.js'},{'revision':null,'url':'./6354.bundle.d8a592b03e9a5b7a66c2.js'},{'revision':null,'url':'./6376.bundle.70707c4b6255e060bb21.js'},{'revision':null,'url':'./6409.bundle.24edbaa283b7c521b8f7.js'},{'revision':null,'url':'./6876.bundle.c934fedea9e764a6dee9.js'},{'revision':null,'url':'./6939.bundle.41fbdef87597b5172ec6.js'},{'revision':null,'url':'./7159.bundle.897f8b37628c4825813a.js'},{'revision':'20a92cf14c2d8be6cd580f68780cc7fa','url':'./7159.css'},{'revision':null,'url':'./7200.bundle.5fa5ec43a8746a209e51.js'},{'revision':null,'url':'./732.bundle.ea6f9d8504e37e01a208.js'},{'revision':null,'url':'./7431.bundle.b31ebb4a2625e89d864e.js'},{'revision':null,'url':'./7639.bundle.2d2d8e9a7a2b7ebb5f0d.js'},{'revision':null,'url':'./7711.bundle.6e7d196e481626c6da19.js'},{'revision':null,'url':'./7739.bundle.284ecb0403a55285c0cb.js'},{'revision':null,'url':'./7758.bundle.c8d106364298e7d288f0.js'},{'revision':null,'url':'./8094.bundle.148a66619607e37cbf19.js'},{'revision':null,'url':'./8271.bundle.a14605302efc3ba38625.js'},{'revision':null,'url':'./8305.bundle.309e576248179fe14315.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./8305.css'},{'revision':null,'url':'./85.bundle.9a8d21329059ff1fd5f8.js'},{'revision':null,'url':'./8558.bundle.23482691a1f00531ff64.js'},{'revision':null,'url':'./8583.bundle.762529c834782843b7cc.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./8583.css'},{'revision':null,'url':'./8665.bundle.dc56c125411422f9f686.js'},{'revision':null,'url':'./9205.bundle.2a2f38a4ee23647f5fa4.js'},{'revision':null,'url':'./9605.bundle.d936cc8cd62e814c9a9f.js'},{'revision':'f8f75c089652cb73f808a2429992e36d','url':'./9605.css'},{'revision':null,'url':'./982b0633e475d94949df.wasm'},{'revision':null,'url':'./9845.bundle.57924e16778260ebad60.js'},{'revision':null,'url':'./9862.bundle.5763cb6044cb4fff7cdd.js'},{'revision':null,'url':'./9927.bundle.0c64f8bd3e753c891eab.js'},{'revision':'d41d8cd98f00b204e9800998ecf8427e','url':'./_headers'},{'revision':'e3bf0f3e9c34f51ad59836ae8e8eaf43','url':'./_redirects'},{'revision':null,'url':'./a957c5cd315f89b6feb1.wasm'},{'revision':'d232c01e1c826849ffc0b23a2b471cfc','url':'./app-config.js'},{'revision':null,'url':'./app.bundle.13cca4f1c7f4b7808e38.js'},{'revision':'76562c39ee1201e311a0fceba1c8695b','url':'./app.bundle.css'},{'revision':'cb4f64534cdf8dd88f1d7219d44490db','url':'./assets/android-chrome-144x144.png'},{'revision':'5cde390de8a619ebe55a669d2ac3effd','url':'./assets/android-chrome-192x192.png'},{'revision':'e7466a67e90471de05401e53b8fe20be','url':'./assets/android-chrome-256x256.png'},{'revision':'9bbe9b80156e930d19a4e1725aa9ddae','url':'./assets/android-chrome-36x36.png'},{'revision':'5698b2ac0c82fe06d84521fc5482df04','url':'./assets/android-chrome-384x384.png'},{'revision':'56bef3fceec344d9747f8abe9c0bba27','url':'./assets/android-chrome-48x48.png'},{'revision':'3e8b8a01290992e82c242557417b0596','url':'./assets/android-chrome-512x512.png'},{'revision':'517925e91e2ce724432d296b687d25e2','url':'./assets/android-chrome-72x72.png'},{'revision':'4c3289bc690f8519012686888e08da71','url':'./assets/android-chrome-96x96.png'},{'revision':'cf464289183184df09292f581df0fb4f','url':'./assets/apple-touch-icon-1024x1024.png'},{'revision':'0857c5282c594e4900e8b31e3bade912','url':'./assets/apple-touch-icon-114x114.png'},{'revision':'4208f41a28130a67e9392a9dfcee6011','url':'./assets/apple-touch-icon-120x120.png'},{'revision':'cb4f64534cdf8dd88f1d7219d44490db','url':'./assets/apple-touch-icon-144x144.png'},{'revision':'977d293982af7e9064ba20806b45cf35','url':'./assets/apple-touch-icon-152x152.png'},{'revision':'6de91b4d2a30600b410758405cb567b4','url':'./assets/apple-touch-icon-167x167.png'},{'revision':'87bff140e3773bd7479a620501c4aa5c','url':'./assets/apple-touch-icon-180x180.png'},{'revision':'647386c34e75f1213830ea9a38913525','url':'./assets/apple-touch-icon-57x57.png'},{'revision':'0c200fe83953738b330ea431083e7a86','url':'./assets/apple-touch-icon-60x60.png'},{'revision':'517925e91e2ce724432d296b687d25e2','url':'./assets/apple-touch-icon-72x72.png'},{'revision':'c9989a807bb18633f6dcf254b5b56124','url':'./assets/apple-touch-icon-76x76.png'},{'revision':'87bff140e3773bd7479a620501c4aa5c','url':'./assets/apple-touch-icon-precomposed.png'},{'revision':'87bff140e3773bd7479a620501c4aa5c','url':'./assets/apple-touch-icon.png'},{'revision':'05fa74ea9c1c0c3931ba96467999081d','url':'./assets/apple-touch-startup-image-1182x2208.png'},{'revision':'9e2cd03e1e6fd0520eea6846f4278018','url':'./assets/apple-touch-startup-image-1242x2148.png'},{'revision':'5591e3a1822cbc8439b99c1a40d53425','url':'./assets/apple-touch-startup-image-1496x2048.png'},{'revision':'337de578c5ca04bd7d2be19d24d83821','url':'./assets/apple-touch-startup-image-1536x2008.png'},{'revision':'cafb4ab4eafe6ef946bd229a1d88e7de','url':'./assets/apple-touch-startup-image-320x460.png'},{'revision':'d9bb9e558d729eeac5efb8be8d6111cc','url':'./assets/apple-touch-startup-image-640x1096.png'},{'revision':'038b5b02bac8b82444bf9a87602ac216','url':'./assets/apple-touch-startup-image-640x920.png'},{'revision':'2177076eb07b1d64d663d7c03268be00','url':'./assets/apple-touch-startup-image-748x1024.png'},{'revision':'4fc097443815fe92503584c4bd73c630','url':'./assets/apple-touch-startup-image-750x1294.png'},{'revision':'2e29914062dce5c5141ab47eea2fc5d9','url':'./assets/apple-touch-startup-image-768x1004.png'},{'revision':'f692ec286b3a332c17985f4ed38b1076','url':'./assets/browserconfig.xml'},{'revision':'f3d9a3b647853c45b0e132e4acd0cc4a','url':'./assets/coast-228x228.png'},{'revision':'ad6e1def5c66193d649a31474bbfe45d','url':'./assets/favicon-16x16.png'},{'revision':'84d1dcdb6cdfa55e2f46be0c80fa5698','url':'./assets/favicon-32x32.png'},{'revision':'95fb44c4998a46109e49d724c060db24','url':'./assets/favicon.ico'},{'revision':'5df2a5b0cee399ac0bc40af74ba3c2cb','url':'./assets/firefox_app_128x128.png'},{'revision':'11fd9098c4b07c8a07e1d2a1e309e046','url':'./assets/firefox_app_512x512.png'},{'revision':'27cddfc922dca3bfa27b4a00fc2f5e36','url':'./assets/firefox_app_60x60.png'},{'revision':'2547879ee2928a5fd56dca483bcbd967','url':'./assets/images/CT-AAA.png'},{'revision':'245628fb9d18d6c5cd529615079f8e86','url':'./assets/images/CT-AAA2.png'},{'revision':'997a9c8b57e7b4e46d8f01fb49d078c6','url':'./assets/images/CT-Air.png'},{'revision':'e7ed89711f63348d49898da3a2b73ab9','url':'./assets/images/CT-Bone.png'},{'revision':'09d1dc6efebb10a60cdf85ed74a10214','url':'./assets/images/CT-Bones.png'},{'revision':'f8ae9bafd3b90ba3efbdcbe8901db187','url':'./assets/images/CT-Cardiac.png'},{'revision':'ed0cc538aba90d6a1c029923ca9b47db','url':'./assets/images/CT-Cardiac2.png'},{'revision':'dfdaf4e4889aefbf09514545f45726bb','url':'./assets/images/CT-Cardiac3.png'},{'revision':'5d4b9a331136a1a5d7ce7ccf7dfcca36','url':'./assets/images/CT-Chest-Contrast-Enhanced.png'},{'revision':'5fb264499a5bb42487663b00e42d1513','url':'./assets/images/CT-Chest-Vessels.png'},{'revision':'ab8f0e569c972388ddaaa61b92449fac','url':'./assets/images/CT-Coronary-Arteries-2.png'},{'revision':'e4e7e2e3cc09bc8c4fb03b6441bd2845','url':'./assets/images/CT-Coronary-Arteries-3.png'},{'revision':'79a2962a2d3cbe4bd3b675b5ca83c659','url':'./assets/images/CT-Coronary-Arteries.png'},{'revision':'6502a6bdab1c70e2188c3bf12ce5d4e4','url':'./assets/images/CT-Cropped-Volume-Bone.png'},{'revision':'e28ad3cf5ed2006fc5a7cffd602736fc','url':'./assets/images/CT-Fat.png'},{'revision':'ca1c5be76053edee27e68ffbf9105101','url':'./assets/images/CT-Liver-Vasculature.png'},{'revision':'d6c9ab857ed2ad19f1ff20bea02b6655','url':'./assets/images/CT-Lung.png'},{'revision':'ddb599c6a163bdf28ece8de1a337470a','url':'./assets/images/CT-MIP.png'},{'revision':'0b0dcf7940a555e2a9a0900f7f84d4d3','url':'./assets/images/CT-Muscle.png'},{'revision':'6082d383f9ce375a5873e2a00fc4d372','url':'./assets/images/CT-Pulmonary-Arteries.png'},{'revision':'b317366149ee2607bec95d63eb4d7661','url':'./assets/images/CT-Soft-Tissue.png'},{'revision':'4b0cd5a673ebb71da22a91ed608d6695','url':'./assets/images/DTI-FA-Brain.png'},{'revision':'3ea1cbee022af5b94710b73726308832','url':'./assets/images/MR-Angio.png'},{'revision':'bd25ca7b166193547f34950eed440e5c','url':'./assets/images/MR-Default.png'},{'revision':'2e50a49b859aeb316a1c23f911fce9ad','url':'./assets/images/MR-MIP.png'},{'revision':'6ba74de6961e08652ff9e2f1871b473d','url':'./assets/images/MR-T2-Brain.png'},{'revision':'42406bb474f0f87cf04a5afbbcd5e389','url':'./assets/images/VolumeRendering.png'},{'revision':'2017d95fae79dcf34b5a5b52586d4763','url':'./assets/manifest.webapp'},{'revision':'cb4f64534cdf8dd88f1d7219d44490db','url':'./assets/mstile-144x144.png'},{'revision':'334895225e16a7777e45d81964725a97','url':'./assets/mstile-150x150.png'},{'revision':'e295cca4af6ed0365cf7b014d91b0e9d','url':'./assets/mstile-310x150.png'},{'revision':'cbefa8c42250e5f2443819fe2c69d91e','url':'./assets/mstile-310x310.png'},{'revision':'aa411a69df2b33a1362fa38d1257fa9d','url':'./assets/mstile-70x70.png'},{'revision':'5609af4f69e40e33471aee770ea1d802','url':'./assets/yandex-browser-50x50.png'},{'revision':'dd001f21b3970d5a7f3e245cc10d21df','url':'./assets/yandex-browser-manifest.json'},{'revision':null,'url':'./c22b37c3488e1d6c3aa4.wasm'},{'revision':null,'url':'./compute.bundle.97539c10dd60d08fbd62.js'},{'revision':'c377e1f5fe4a207d270c3f7a8dd3e3ca','url':'./dicom-microscopy-viewer/charlswasm_decode.wasm'},{'revision':'9214592a1edd456ad2ef69772201398e','url':'./dicom-microscopy-viewer/dataLoader.worker.min.js'},{'revision':'813dee9adb3a3478dedc6dfe1479e15b','url':'./dicom-microscopy-viewer/dataLoader.worker.min.js.map'},{'revision':'fb088ddc5ec5de63aa6095f70baa2d45','url':'./dicom-microscopy-viewer/dicomMicroscopyViewer.min.js'},{'revision':'cdf943673c34e8dd6be85b026bdffc70','url':'./dicom-microscopy-viewer/dicomMicroscopyViewer.min.js.map'},{'revision':'52b9a07fe0541fe8c313d9788550bf51','url':'./dicom-microscopy-viewer/dicomiccwasm.wasm'},{'revision':'7edb59d2be7c993050cb31ded36afa31','url':'./dicom-microscopy-viewer/libjpegturbowasm_decode.wasm'},{'revision':'54a25d82a224dfd401569baffc8b478d','url':'./dicom-microscopy-viewer/openjpegwasm_decode.wasm'},{'revision':'4acdd19a35d759ec2669f1ba9490937d','url':'./es6-shim.min.js'},{'revision':'92b01d0a387d59841e02cc3a4c994657','url':'./google.js'},{'revision':null,'url':'./histogram-worker.bundle.96d0f655b219f175d38c.js'},{'revision':'a29ebcbe0597116395ca058f57c6f027','url':'./index.html'},{'revision':'31c0367ca4160b2c6373e905739c5719','url':'./init-service-worker.js'},{'revision':null,'url':'./interpolation.bundle.66ff78ab4e086e262118.js'},{'revision':'74fc9658b62903be2048c1f82a22b4d4','url':'./manifest.json'},{'revision':'3fa71aa0af3e34b4ebd9a71eee0f4bdd','url':'./ohif-logo-light.svg'},{'revision':'7e81da785c63e75650101db6c5d7560e','url':'./ohif-logo.svg'},{'revision':'eadf8bf1d85032a029e2c0df4b8938b0','url':'./oidc-client.min.js'},{'revision':'242abae066a3d4717e09f218703f1923','url':'./ort/cjs/ort.all.min.js'},{'revision':'039ea411b3b655094035e85bb2e245d4','url':'./ort/cjs/ort.all.min.js.map'},{'revision':'fb509ab752acf788c3cc25459ae72e62','url':'./ort/cjs/ort.min.js'},{'revision':'30087ad61fc31272eb6094962e3a0451','url':'./ort/cjs/ort.min.js.map'},{'revision':'0fad6d8ca65291350ca8a1751659a536','url':'./ort/cjs/ort.training.wasm.min.js'},{'revision':'da692fb2cd7f67223b92a90955b87cb3','url':'./ort/cjs/ort.training.wasm.min.js.map'},{'revision':'f471b605fb03df54c63d5bc77a32226b','url':'./ort/cjs/ort.wasm-core.min.js'},{'revision':'cc4475faebff69233b4fdc1a78eb0399','url':'./ort/cjs/ort.wasm-core.min.js.map'},{'revision':'d228221765171d0a4cf3cb3909dd0be3','url':'./ort/cjs/ort.wasm.min.js'},{'revision':'6d132c02e5eeadf4da54c8bcc0c15200','url':'./ort/cjs/ort.wasm.min.js.map'},{'revision':'9a463e37f9550d7a23f5a47076b6b3d8','url':'./ort/cjs/ort.webgl.min.js'},{'revision':'26e251b6aee08c91a3d6e1af1410659d','url':'./ort/cjs/ort.webgl.min.js.map'},{'revision':'5cd623012201d7c5ddbbdb57224ac26d','url':'./ort/cjs/ort.webgpu.min.js'},{'revision':'e53c54f9663a5d7573b66dd3d25fb93a','url':'./ort/cjs/ort.webgpu.min.js.map'},{'revision':'a6b7eb31b5dcf3835a90ea0d201f21e3','url':'./ort/cjs/package.json'},{'revision':'2ccbe0c4397149108dd0d7ced00b5879','url':'./ort/esm/ort.all.min.js'},{'revision':'2b2ff793ba11a59563a504067aa1e469','url':'./ort/esm/ort.all.min.js.map'},{'revision':'978e978337e66e848c2351da61fd7e8e','url':'./ort/esm/ort.min.js'},{'revision':'add9f88f20194b214c8a97b200c4eac0','url':'./ort/esm/ort.min.js.map'},{'revision':'61471773b94ac06d0d814b49cf619233','url':'./ort/esm/ort.training.wasm.min.js'},{'revision':'60057feb2e9c1d39d7e52b165e450e72','url':'./ort/esm/ort.training.wasm.min.js.map'},{'revision':'12635093e14a67921565de89f678c470','url':'./ort/esm/ort.wasm-core.min.js'},{'revision':'1d2c905465e8b47dc023beaf715bb6ae','url':'./ort/esm/ort.wasm-core.min.js.map'},{'revision':'f2780d73ea108c5728b4c4fb6bd1e8df','url':'./ort/esm/ort.wasm.min.js'},{'revision':'e1af0f0e8befaa973a9192cdc1a882fb','url':'./ort/esm/ort.wasm.min.js.map'},{'revision':'c5a2e8ed939ccfb0211336d8b385d81a','url':'./ort/esm/ort.webgl.min.js'},{'revision':'9e0198bc10517578066aed6c8baef1a3','url':'./ort/esm/ort.webgl.min.js.map'},{'revision':'0c96a22438d8ffdf94f13b1693dbed8b','url':'./ort/esm/ort.webgpu.min.js'},{'revision':'24126f825e1c2dd678b4b85cde6c4e3a','url':'./ort/esm/ort.webgpu.min.js.map'},{'revision':'b305354b5d3207dc798e87449005a1de','url':'./ort/esm/package.json'},{'revision':'10ba4d1a5289405cff7d9b5f453b60b4','url':'./ort/ort-training-wasm-simd.wasm'},{'revision':'f7380fb5928344460bac0e5833ee4a82','url':'./ort/ort-wasm-simd-threaded.jsep.wasm'},{'revision':'068e15f5aba004197782c011a782c500','url':'./ort/ort-wasm-simd-threaded.wasm'},{'revision':'8a468bda81bfb9e375907f0b6da8ffee','url':'./ort/ort-wasm-simd.jsep.wasm'},{'revision':'27a8292511e77fc50ced64326875ca4a','url':'./ort/ort-wasm-simd.wasm'},{'revision':'c9cf2d09cfabe6ea517c87620091a2fb','url':'./ort/ort-wasm-threaded.wasm'},{'revision':'88cb0475e8c4228b29dd201646223f0c','url':'./ort/ort-wasm.wasm'},{'revision':'708def774c211a2b036384d9e14dbfe7','url':'./ort/ort.all.js'},{'revision':'32e292212f5767983e5433f0a88d3425','url':'./ort/ort.all.min.js'},{'revision':'c4a4b66af400491c83591ab732dac868','url':'./ort/ort.all.min.js.map'},{'revision':'5d78ec66bcab8a26756d43b8c63a8f1e','url':'./ort/ort.js'},{'revision':'d552a78742409520bd30fb824a88a266','url':'./ort/ort.min.js'},{'revision':'a1fdb4abef4a6130f72148fe469ca114','url':'./ort/ort.min.js.map'},{'revision':'e3b420cb98f192d8755934dde7c014db','url':'./ort/ort.node.min.js'},{'revision':'9b0c2098e6c9b0f6d6e54e8cb434ed76','url':'./ort/ort.node.min.js.map'},{'revision':'722dabcfbb18772ca80ef96ef9b9a036','url':'./ort/ort.training.wasm.min.js'},{'revision':'de35b3d2e302163bdcb80d3e20c84266','url':'./ort/ort.training.wasm.min.js.map'},{'revision':'f92b62229fc36d346f720b2410e63576','url':'./ort/ort.wasm-core.min.js'},{'revision':'8f56640956155afecadc318d8a32e0f6','url':'./ort/ort.wasm-core.min.js.map'},{'revision':'5ac1ffe964e3bd1f0d2e6a3a18a5a1ba','url':'./ort/ort.wasm.min.js'},{'revision':'9b0a74e48529a07b61428b1a48233b39','url':'./ort/ort.wasm.min.js.map'},{'revision':'4591c9c641ab0151b9a8cdc1ce110286','url':'./ort/ort.webgl.min.js'},{'revision':'77e632e10b98c8d10e17bb3071d5cba4','url':'./ort/ort.webgl.min.js.map'},{'revision':'027aca68552f5081e90a08f7fb064fb7','url':'./ort/ort.webgpu.min.js'},{'revision':'12dc9d5c577bac166f0f74590e4d27dd','url':'./ort/ort.webgpu.min.js.map'},{'revision':null,'url':'./polySeg.bundle.e377acbd164c676dd471.js'},{'revision':'a1aef5311245f5864315443d12246c37','url':'./polyfill.min.js'},{'revision':'981ac7f7543b8330e3826aff72c33382','url':'./serve.json'},{'revision':'b1e488d9955b62bd2858874df11d5223','url':'./silent-refresh.html'}]);
+workbox.precaching.precacheAndRoute([{'revision':null,'url':'./1398.bundle.667fbe38ca6d9fdfea19.js'},{'revision':null,'url':'./1403.bundle.75d8ae601fad37ac59b9.js'},{'revision':null,'url':'./1459.bundle.0df00d11af8704537c76.js'},{'revision':null,'url':'./147.bundle.037f190998936d61183b.js'},{'revision':'e0cfcadcf2c2fbc1f117196b9f62bf3e','url':'./147.css'},{'revision':null,'url':'./1554.bundle.313701fc6b7c2b153c28.js'},{'revision':null,'url':'./1608.bundle.7879b1f7ca74b87faf8c.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./1608.css'},{'revision':null,'url':'./1730.bundle.9a528d67a7ddfefd4ac8.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./1730.css'},{'revision':null,'url':'./1758.bundle.9574b448d811268c73b4.js'},{'revision':null,'url':'./17dd54813d5acc10bf8f.wasm'},{'revision':null,'url':'./1927.bundle.da05087c08bf79ab6164.js'},{'revision':null,'url':'./1933.bundle.f47a3916c63908ce4510.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./1933.css'},{'revision':null,'url':'./2018.bundle.a75601d36f8335e6ad70.js'},{'revision':null,'url':'./2068.bundle.bbd5a91f3a9f27e42002.js'},{'revision':null,'url':'./213.bundle.4f5a9f54de1d8b00dbe0.js'},{'revision':null,'url':'./2199.bundle.05379e24d0559fb1b62d.js'},{'revision':null,'url':'./2424.bundle.200b0ae6a4293279fd9b.js'},{'revision':null,'url':'./2516.bundle.299b43d001321aeb1554.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./2591.css'},{'revision':null,'url':'./2701.bundle.57fa57f3211b5af30333.js'},{'revision':null,'url':'./2764.bundle.eeb852510f3006df2e59.js'},{'revision':null,'url':'./2816.bundle.838fa4d72921c026083c.js'},{'revision':null,'url':'./2838.bundle.8d4f0cf2c02681ca6878.js'},{'revision':null,'url':'./31fb9346313fc3740d7b.woff2'},{'revision':null,'url':'./3343.bundle.a4424b8418b8c0c1a9e4.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./3343.css'},{'revision':null,'url':'./3461.bundle.58547d3ab516df762f48.js'},{'revision':null,'url':'./3562.bundle.6b7411fbb94c93a34820.js'},{'revision':null,'url':'./3910.bundle.b6180cdfa45c750c206b.js'},{'revision':'2b8874726752c2c40347df5b55a1e4f9','url':'./3910.css'},{'revision':null,'url':'./3933.bundle.aadc7dabcb74e4ca266c.js'},{'revision':'afd04227d29073ce2d3c1662140270d7','url':'./3933.css'},{'revision':null,'url':'./4202.bundle.be22ef13c3844ef0f8c4.js'},{'revision':null,'url':'./4287.bundle.87e14e648f2bce5e4176.js'},{'revision':null,'url':'./4775.bundle.d0ea4b37875cfa85e56c.js'},{'revision':null,'url':'./4819.bundle.f7757d944c785f953b6b.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./4819.css'},{'revision':null,'url':'./5004fdc02f329ce53b69.wasm'},{'revision':null,'url':'./5028.bundle.cd59aeb28444f14ad1ee.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./5028.css'},{'revision':null,'url':'./5261.bundle.2655560097e9250eac44.js'},{'revision':null,'url':'./5299.bundle.4ae3f0d5a95b5483f5e9.js'},{'revision':null,'url':'./5457.bundle.c44221459be9b9c5937c.js'},{'revision':null,'url':'./5485.bundle.1add1e84bf0652e1f978.js'},{'revision':null,'url':'./5491.bundle.4866d2ecb20dd089e071.js'},{'revision':null,'url':'./5547.bundle.10a9625eac35067f6c77.js'},{'revision':'7b35665742c835684bc86d3fa5397b30','url':'./5547.css'},{'revision':null,'url':'./5609.bundle.41430e142f0b95cb32e9.js'},{'revision':null,'url':'./5724.bundle.62ea5c07c7809ab4cd74.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./5724.css'},{'revision':null,'url':'./5802.bundle.d0b1eda248c9ce5ff880.js'},{'revision':'78c6e381862b3f33316bcffadde4ccd6','url':'./5802.css'},{'revision':null,'url':'./5830.bundle.50dcea7ba7954661d2b7.js'},{'revision':null,'url':'./5858.bundle.466e58128de344ab53f3.js'},{'revision':null,'url':'./6027.bundle.4b70caf6273a7926016d.js'},{'revision':null,'url':'./6163.bundle.09921650f5c2a777c59a.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./6163.css'},{'revision':null,'url':'./6280.bundle.f22f9955eb906e1d7da8.js'},{'revision':null,'url':'./6347.bundle.46b118d368710e08c546.js'},{'revision':null,'url':'./6354.bundle.d8a592b03e9a5b7a66c2.js'},{'revision':null,'url':'./6376.bundle.70707c4b6255e060bb21.js'},{'revision':null,'url':'./6409.bundle.24edbaa283b7c521b8f7.js'},{'revision':null,'url':'./6876.bundle.c934fedea9e764a6dee9.js'},{'revision':null,'url':'./6939.bundle.41fbdef87597b5172ec6.js'},{'revision':null,'url':'./7159.bundle.897f8b37628c4825813a.js'},{'revision':'20a92cf14c2d8be6cd580f68780cc7fa','url':'./7159.css'},{'revision':null,'url':'./7200.bundle.5fa5ec43a8746a209e51.js'},{'revision':null,'url':'./732.bundle.ea6f9d8504e37e01a208.js'},{'revision':null,'url':'./7431.bundle.b31ebb4a2625e89d864e.js'},{'revision':null,'url':'./7639.bundle.2d2d8e9a7a2b7ebb5f0d.js'},{'revision':null,'url':'./7711.bundle.6e7d196e481626c6da19.js'},{'revision':null,'url':'./7739.bundle.284ecb0403a55285c0cb.js'},{'revision':null,'url':'./7758.bundle.c8d106364298e7d288f0.js'},{'revision':null,'url':'./8094.bundle.148a66619607e37cbf19.js'},{'revision':null,'url':'./8271.bundle.a14605302efc3ba38625.js'},{'revision':null,'url':'./8305.bundle.309e576248179fe14315.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./8305.css'},{'revision':null,'url':'./85.bundle.9a8d21329059ff1fd5f8.js'},{'revision':null,'url':'./8558.bundle.23482691a1f00531ff64.js'},{'revision':null,'url':'./8583.bundle.762529c834782843b7cc.js'},{'revision':'27c847594562926775c007c401b4fe98','url':'./8583.css'},{'revision':null,'url':'./8665.bundle.dc56c125411422f9f686.js'},{'revision':null,'url':'./9205.bundle.2a2f38a4ee23647f5fa4.js'},{'revision':null,'url':'./9605.bundle.ff0b5c0ab0221f89ba28.js'},{'revision':'f8f75c089652cb73f808a2429992e36d','url':'./9605.css'},{'revision':null,'url':'./982b0633e475d94949df.wasm'},{'revision':null,'url':'./9845.bundle.57924e16778260ebad60.js'},{'revision':null,'url':'./9862.bundle.5763cb6044cb4fff7cdd.js'},{'revision':null,'url':'./9927.bundle.0c64f8bd3e753c891eab.js'},{'revision':'d41d8cd98f00b204e9800998ecf8427e','url':'./_headers'},{'revision':'e3bf0f3e9c34f51ad59836ae8e8eaf43','url':'./_redirects'},{'revision':null,'url':'./a957c5cd315f89b6feb1.wasm'},{'revision':'d232c01e1c826849ffc0b23a2b471cfc','url':'./app-config.js'},{'revision':null,'url':'./app.bundle.6bf41afda50f7ddf6348.js'},{'revision':'76562c39ee1201e311a0fceba1c8695b','url':'./app.bundle.css'},{'revision':'cb4f64534cdf8dd88f1d7219d44490db','url':'./assets/android-chrome-144x144.png'},{'revision':'5cde390de8a619ebe55a669d2ac3effd','url':'./assets/android-chrome-192x192.png'},{'revision':'e7466a67e90471de05401e53b8fe20be','url':'./assets/android-chrome-256x256.png'},{'revision':'9bbe9b80156e930d19a4e1725aa9ddae','url':'./assets/android-chrome-36x36.png'},{'revision':'5698b2ac0c82fe06d84521fc5482df04','url':'./assets/android-chrome-384x384.png'},{'revision':'56bef3fceec344d9747f8abe9c0bba27','url':'./assets/android-chrome-48x48.png'},{'revision':'3e8b8a01290992e82c242557417b0596','url':'./assets/android-chrome-512x512.png'},{'revision':'517925e91e2ce724432d296b687d25e2','url':'./assets/android-chrome-72x72.png'},{'revision':'4c3289bc690f8519012686888e08da71','url':'./assets/android-chrome-96x96.png'},{'revision':'cf464289183184df09292f581df0fb4f','url':'./assets/apple-touch-icon-1024x1024.png'},{'revision':'0857c5282c594e4900e8b31e3bade912','url':'./assets/apple-touch-icon-114x114.png'},{'revision':'4208f41a28130a67e9392a9dfcee6011','url':'./assets/apple-touch-icon-120x120.png'},{'revision':'cb4f64534cdf8dd88f1d7219d44490db','url':'./assets/apple-touch-icon-144x144.png'},{'revision':'977d293982af7e9064ba20806b45cf35','url':'./assets/apple-touch-icon-152x152.png'},{'revision':'6de91b4d2a30600b410758405cb567b4','url':'./assets/apple-touch-icon-167x167.png'},{'revision':'87bff140e3773bd7479a620501c4aa5c','url':'./assets/apple-touch-icon-180x180.png'},{'revision':'647386c34e75f1213830ea9a38913525','url':'./assets/apple-touch-icon-57x57.png'},{'revision':'0c200fe83953738b330ea431083e7a86','url':'./assets/apple-touch-icon-60x60.png'},{'revision':'517925e91e2ce724432d296b687d25e2','url':'./assets/apple-touch-icon-72x72.png'},{'revision':'c9989a807bb18633f6dcf254b5b56124','url':'./assets/apple-touch-icon-76x76.png'},{'revision':'87bff140e3773bd7479a620501c4aa5c','url':'./assets/apple-touch-icon-precomposed.png'},{'revision':'87bff140e3773bd7479a620501c4aa5c','url':'./assets/apple-touch-icon.png'},{'revision':'05fa74ea9c1c0c3931ba96467999081d','url':'./assets/apple-touch-startup-image-1182x2208.png'},{'revision':'9e2cd03e1e6fd0520eea6846f4278018','url':'./assets/apple-touch-startup-image-1242x2148.png'},{'revision':'5591e3a1822cbc8439b99c1a40d53425','url':'./assets/apple-touch-startup-image-1496x2048.png'},{'revision':'337de578c5ca04bd7d2be19d24d83821','url':'./assets/apple-touch-startup-image-1536x2008.png'},{'revision':'cafb4ab4eafe6ef946bd229a1d88e7de','url':'./assets/apple-touch-startup-image-320x460.png'},{'revision':'d9bb9e558d729eeac5efb8be8d6111cc','url':'./assets/apple-touch-startup-image-640x1096.png'},{'revision':'038b5b02bac8b82444bf9a87602ac216','url':'./assets/apple-touch-startup-image-640x920.png'},{'revision':'2177076eb07b1d64d663d7c03268be00','url':'./assets/apple-touch-startup-image-748x1024.png'},{'revision':'4fc097443815fe92503584c4bd73c630','url':'./assets/apple-touch-startup-image-750x1294.png'},{'revision':'2e29914062dce5c5141ab47eea2fc5d9','url':'./assets/apple-touch-startup-image-768x1004.png'},{'revision':'f692ec286b3a332c17985f4ed38b1076','url':'./assets/browserconfig.xml'},{'revision':'f3d9a3b647853c45b0e132e4acd0cc4a','url':'./assets/coast-228x228.png'},{'revision':'ad6e1def5c66193d649a31474bbfe45d','url':'./assets/favicon-16x16.png'},{'revision':'84d1dcdb6cdfa55e2f46be0c80fa5698','url':'./assets/favicon-32x32.png'},{'revision':'95fb44c4998a46109e49d724c060db24','url':'./assets/favicon.ico'},{'revision':'5df2a5b0cee399ac0bc40af74ba3c2cb','url':'./assets/firefox_app_128x128.png'},{'revision':'11fd9098c4b07c8a07e1d2a1e309e046','url':'./assets/firefox_app_512x512.png'},{'revision':'27cddfc922dca3bfa27b4a00fc2f5e36','url':'./assets/firefox_app_60x60.png'},{'revision':'2547879ee2928a5fd56dca483bcbd967','url':'./assets/images/CT-AAA.png'},{'revision':'245628fb9d18d6c5cd529615079f8e86','url':'./assets/images/CT-AAA2.png'},{'revision':'997a9c8b57e7b4e46d8f01fb49d078c6','url':'./assets/images/CT-Air.png'},{'revision':'e7ed89711f63348d49898da3a2b73ab9','url':'./assets/images/CT-Bone.png'},{'revision':'09d1dc6efebb10a60cdf85ed74a10214','url':'./assets/images/CT-Bones.png'},{'revision':'f8ae9bafd3b90ba3efbdcbe8901db187','url':'./assets/images/CT-Cardiac.png'},{'revision':'ed0cc538aba90d6a1c029923ca9b47db','url':'./assets/images/CT-Cardiac2.png'},{'revision':'dfdaf4e4889aefbf09514545f45726bb','url':'./assets/images/CT-Cardiac3.png'},{'revision':'5d4b9a331136a1a5d7ce7ccf7dfcca36','url':'./assets/images/CT-Chest-Contrast-Enhanced.png'},{'revision':'5fb264499a5bb42487663b00e42d1513','url':'./assets/images/CT-Chest-Vessels.png'},{'revision':'ab8f0e569c972388ddaaa61b92449fac','url':'./assets/images/CT-Coronary-Arteries-2.png'},{'revision':'e4e7e2e3cc09bc8c4fb03b6441bd2845','url':'./assets/images/CT-Coronary-Arteries-3.png'},{'revision':'79a2962a2d3cbe4bd3b675b5ca83c659','url':'./assets/images/CT-Coronary-Arteries.png'},{'revision':'6502a6bdab1c70e2188c3bf12ce5d4e4','url':'./assets/images/CT-Cropped-Volume-Bone.png'},{'revision':'e28ad3cf5ed2006fc5a7cffd602736fc','url':'./assets/images/CT-Fat.png'},{'revision':'ca1c5be76053edee27e68ffbf9105101','url':'./assets/images/CT-Liver-Vasculature.png'},{'revision':'d6c9ab857ed2ad19f1ff20bea02b6655','url':'./assets/images/CT-Lung.png'},{'revision':'ddb599c6a163bdf28ece8de1a337470a','url':'./assets/images/CT-MIP.png'},{'revision':'0b0dcf7940a555e2a9a0900f7f84d4d3','url':'./assets/images/CT-Muscle.png'},{'revision':'6082d383f9ce375a5873e2a00fc4d372','url':'./assets/images/CT-Pulmonary-Arteries.png'},{'revision':'b317366149ee2607bec95d63eb4d7661','url':'./assets/images/CT-Soft-Tissue.png'},{'revision':'4b0cd5a673ebb71da22a91ed608d6695','url':'./assets/images/DTI-FA-Brain.png'},{'revision':'3ea1cbee022af5b94710b73726308832','url':'./assets/images/MR-Angio.png'},{'revision':'bd25ca7b166193547f34950eed440e5c','url':'./assets/images/MR-Default.png'},{'revision':'2e50a49b859aeb316a1c23f911fce9ad','url':'./assets/images/MR-MIP.png'},{'revision':'6ba74de6961e08652ff9e2f1871b473d','url':'./assets/images/MR-T2-Brain.png'},{'revision':'42406bb474f0f87cf04a5afbbcd5e389','url':'./assets/images/VolumeRendering.png'},{'revision':'2017d95fae79dcf34b5a5b52586d4763','url':'./assets/manifest.webapp'},{'revision':'cb4f64534cdf8dd88f1d7219d44490db','url':'./assets/mstile-144x144.png'},{'revision':'334895225e16a7777e45d81964725a97','url':'./assets/mstile-150x150.png'},{'revision':'e295cca4af6ed0365cf7b014d91b0e9d','url':'./assets/mstile-310x150.png'},{'revision':'cbefa8c42250e5f2443819fe2c69d91e','url':'./assets/mstile-310x310.png'},{'revision':'aa411a69df2b33a1362fa38d1257fa9d','url':'./assets/mstile-70x70.png'},{'revision':'5609af4f69e40e33471aee770ea1d802','url':'./assets/yandex-browser-50x50.png'},{'revision':'dd001f21b3970d5a7f3e245cc10d21df','url':'./assets/yandex-browser-manifest.json'},{'revision':null,'url':'./c22b37c3488e1d6c3aa4.wasm'},{'revision':null,'url':'./compute.bundle.97539c10dd60d08fbd62.js'},{'revision':'c377e1f5fe4a207d270c3f7a8dd3e3ca','url':'./dicom-microscopy-viewer/charlswasm_decode.wasm'},{'revision':'9214592a1edd456ad2ef69772201398e','url':'./dicom-microscopy-viewer/dataLoader.worker.min.js'},{'revision':'813dee9adb3a3478dedc6dfe1479e15b','url':'./dicom-microscopy-viewer/dataLoader.worker.min.js.map'},{'revision':'fb088ddc5ec5de63aa6095f70baa2d45','url':'./dicom-microscopy-viewer/dicomMicroscopyViewer.min.js'},{'revision':'cdf943673c34e8dd6be85b026bdffc70','url':'./dicom-microscopy-viewer/dicomMicroscopyViewer.min.js.map'},{'revision':'52b9a07fe0541fe8c313d9788550bf51','url':'./dicom-microscopy-viewer/dicomiccwasm.wasm'},{'revision':'7edb59d2be7c993050cb31ded36afa31','url':'./dicom-microscopy-viewer/libjpegturbowasm_decode.wasm'},{'revision':'54a25d82a224dfd401569baffc8b478d','url':'./dicom-microscopy-viewer/openjpegwasm_decode.wasm'},{'revision':'4acdd19a35d759ec2669f1ba9490937d','url':'./es6-shim.min.js'},{'revision':'92b01d0a387d59841e02cc3a4c994657','url':'./google.js'},{'revision':null,'url':'./histogram-worker.bundle.96d0f655b219f175d38c.js'},{'revision':'6a490bf7a747acc80ef7cc81afdebafc','url':'./index.html'},{'revision':'31c0367ca4160b2c6373e905739c5719','url':'./init-service-worker.js'},{'revision':null,'url':'./interpolation.bundle.66ff78ab4e086e262118.js'},{'revision':'74fc9658b62903be2048c1f82a22b4d4','url':'./manifest.json'},{'revision':'3fa71aa0af3e34b4ebd9a71eee0f4bdd','url':'./ohif-logo-light.svg'},{'revision':'7e81da785c63e75650101db6c5d7560e','url':'./ohif-logo.svg'},{'revision':'eadf8bf1d85032a029e2c0df4b8938b0','url':'./oidc-client.min.js'},{'revision':'242abae066a3d4717e09f218703f1923','url':'./ort/cjs/ort.all.min.js'},{'revision':'039ea411b3b655094035e85bb2e245d4','url':'./ort/cjs/ort.all.min.js.map'},{'revision':'fb509ab752acf788c3cc25459ae72e62','url':'./ort/cjs/ort.min.js'},{'revision':'30087ad61fc31272eb6094962e3a0451','url':'./ort/cjs/ort.min.js.map'},{'revision':'0fad6d8ca65291350ca8a1751659a536','url':'./ort/cjs/ort.training.wasm.min.js'},{'revision':'da692fb2cd7f67223b92a90955b87cb3','url':'./ort/cjs/ort.training.wasm.min.js.map'},{'revision':'f471b605fb03df54c63d5bc77a32226b','url':'./ort/cjs/ort.wasm-core.min.js'},{'revision':'cc4475faebff69233b4fdc1a78eb0399','url':'./ort/cjs/ort.wasm-core.min.js.map'},{'revision':'d228221765171d0a4cf3cb3909dd0be3','url':'./ort/cjs/ort.wasm.min.js'},{'revision':'6d132c02e5eeadf4da54c8bcc0c15200','url':'./ort/cjs/ort.wasm.min.js.map'},{'revision':'9a463e37f9550d7a23f5a47076b6b3d8','url':'./ort/cjs/ort.webgl.min.js'},{'revision':'26e251b6aee08c91a3d6e1af1410659d','url':'./ort/cjs/ort.webgl.min.js.map'},{'revision':'5cd623012201d7c5ddbbdb57224ac26d','url':'./ort/cjs/ort.webgpu.min.js'},{'revision':'e53c54f9663a5d7573b66dd3d25fb93a','url':'./ort/cjs/ort.webgpu.min.js.map'},{'revision':'a6b7eb31b5dcf3835a90ea0d201f21e3','url':'./ort/cjs/package.json'},{'revision':'2ccbe0c4397149108dd0d7ced00b5879','url':'./ort/esm/ort.all.min.js'},{'revision':'2b2ff793ba11a59563a504067aa1e469','url':'./ort/esm/ort.all.min.js.map'},{'revision':'978e978337e66e848c2351da61fd7e8e','url':'./ort/esm/ort.min.js'},{'revision':'add9f88f20194b214c8a97b200c4eac0','url':'./ort/esm/ort.min.js.map'},{'revision':'61471773b94ac06d0d814b49cf619233','url':'./ort/esm/ort.training.wasm.min.js'},{'revision':'60057feb2e9c1d39d7e52b165e450e72','url':'./ort/esm/ort.training.wasm.min.js.map'},{'revision':'12635093e14a67921565de89f678c470','url':'./ort/esm/ort.wasm-core.min.js'},{'revision':'1d2c905465e8b47dc023beaf715bb6ae','url':'./ort/esm/ort.wasm-core.min.js.map'},{'revision':'f2780d73ea108c5728b4c4fb6bd1e8df','url':'./ort/esm/ort.wasm.min.js'},{'revision':'e1af0f0e8befaa973a9192cdc1a882fb','url':'./ort/esm/ort.wasm.min.js.map'},{'revision':'c5a2e8ed939ccfb0211336d8b385d81a','url':'./ort/esm/ort.webgl.min.js'},{'revision':'9e0198bc10517578066aed6c8baef1a3','url':'./ort/esm/ort.webgl.min.js.map'},{'revision':'0c96a22438d8ffdf94f13b1693dbed8b','url':'./ort/esm/ort.webgpu.min.js'},{'revision':'24126f825e1c2dd678b4b85cde6c4e3a','url':'./ort/esm/ort.webgpu.min.js.map'},{'revision':'b305354b5d3207dc798e87449005a1de','url':'./ort/esm/package.json'},{'revision':'10ba4d1a5289405cff7d9b5f453b60b4','url':'./ort/ort-training-wasm-simd.wasm'},{'revision':'f7380fb5928344460bac0e5833ee4a82','url':'./ort/ort-wasm-simd-threaded.jsep.wasm'},{'revision':'068e15f5aba004197782c011a782c500','url':'./ort/ort-wasm-simd-threaded.wasm'},{'revision':'8a468bda81bfb9e375907f0b6da8ffee','url':'./ort/ort-wasm-simd.jsep.wasm'},{'revision':'27a8292511e77fc50ced64326875ca4a','url':'./ort/ort-wasm-simd.wasm'},{'revision':'c9cf2d09cfabe6ea517c87620091a2fb','url':'./ort/ort-wasm-threaded.wasm'},{'revision':'88cb0475e8c4228b29dd201646223f0c','url':'./ort/ort-wasm.wasm'},{'revision':'708def774c211a2b036384d9e14dbfe7','url':'./ort/ort.all.js'},{'revision':'32e292212f5767983e5433f0a88d3425','url':'./ort/ort.all.min.js'},{'revision':'c4a4b66af400491c83591ab732dac868','url':'./ort/ort.all.min.js.map'},{'revision':'5d78ec66bcab8a26756d43b8c63a8f1e','url':'./ort/ort.js'},{'revision':'d552a78742409520bd30fb824a88a266','url':'./ort/ort.min.js'},{'revision':'a1fdb4abef4a6130f72148fe469ca114','url':'./ort/ort.min.js.map'},{'revision':'e3b420cb98f192d8755934dde7c014db','url':'./ort/ort.node.min.js'},{'revision':'9b0c2098e6c9b0f6d6e54e8cb434ed76','url':'./ort/ort.node.min.js.map'},{'revision':'722dabcfbb18772ca80ef96ef9b9a036','url':'./ort/ort.training.wasm.min.js'},{'revision':'de35b3d2e302163bdcb80d3e20c84266','url':'./ort/ort.training.wasm.min.js.map'},{'revision':'f92b62229fc36d346f720b2410e63576','url':'./ort/ort.wasm-core.min.js'},{'revision':'8f56640956155afecadc318d8a32e0f6','url':'./ort/ort.wasm-core.min.js.map'},{'revision':'5ac1ffe964e3bd1f0d2e6a3a18a5a1ba','url':'./ort/ort.wasm.min.js'},{'revision':'9b0a74e48529a07b61428b1a48233b39','url':'./ort/ort.wasm.min.js.map'},{'revision':'4591c9c641ab0151b9a8cdc1ce110286','url':'./ort/ort.webgl.min.js'},{'revision':'77e632e10b98c8d10e17bb3071d5cba4','url':'./ort/ort.webgl.min.js.map'},{'revision':'027aca68552f5081e90a08f7fb064fb7','url':'./ort/ort.webgpu.min.js'},{'revision':'12dc9d5c577bac166f0f74590e4d27dd','url':'./ort/ort.webgpu.min.js.map'},{'revision':null,'url':'./polySeg.bundle.e377acbd164c676dd471.js'},{'revision':'a1aef5311245f5864315443d12246c37','url':'./polyfill.min.js'},{'revision':'981ac7f7543b8330e3826aff72c33382','url':'./serve.json'},{'revision':'b1e488d9955b62bd2858874df11d5223','url':'./silent-refresh.html'}]);
 
 // TODO: Cache API
 // https://developers.google.com/web/fundamentals/instant-and-offline/web-storage/cache-api
