+++ src/dbx/pixels/resources/UI_VERSION
@@ -1,2 +1,2 @@
-1.1.0
+1.2.0
